Optimize graph storage: SFC node ordering, jagged adj list, PCO edges
- Replace Vec<Vec<(u32,f32)>> adjacency list with JaggedArray<(u32,f32)>
  for flat memory layout and better cache locality during Dijkstra

- Binary format v4 (write + read):
  - Sort nodes by Morton (Z-order) space-filling curve so geographically
    nearby nodes get nearby indices; store lats/lons as two PCO-compressed
    f64 arrays (consecutive values are similar -> good delta compression)
  - Represent each undirected edge canonically with u > v, encode as
    (u, delta=u-v) and sort by (u, delta); store as three PCO-compressed
    u32 arrays (u, delta, dist_bits); small deltas exploit SFC locality
  - Relabel all node references (edges, stop_to_node) to new SFC indices

+++ transit-router/src/router.rs
@@ -219,7 +219,7 @@ fn run_tdd_inner(
         // Walking edges — leave_home and current_event propagate unchanged.
         // Propagating current_event means: if you were mid-trip and walk to an
         // adjacent node, you can still continue the same trip from there.
-        for &(neighbor, distance) in &data.adj[node as usize] {
+        for &(neighbor, distance) in &data.adj[node] {
             let wt = (distance / WALKING_SPEED_MPS) as u32;
             let arrival = t_current + wt;
             let candidate = NodeResult {
